move cuda_characterize, but break it
