+++ src/device/cuda/characterize/math.rs
