+++ src/device/cuda/characterize/table.rs
