+++ src/device/cuda/mod.rs
@@ -14,12 +14,13 @@ mod gpu;
 mod mem_model;
 mod printer;
 
+pub mod characterize;
+
 
 // Constructs to retrieve information on the GPU, that are not needed for the regular
 // operation of Telamon and thus only present if the cuda feature is.
 #[cfg(feature="cuda")]
 pub use self::api::{DeviceAttribute, PerfCounter, PerfCounterSet};
-
 pub use self::api::{Array, Executor, JITDaemon};
 pub use self::context::Context;
 pub use self::gpu::{Gpu, InstDesc};
